Adds utility function
